# Perf: optimize `norm_l1` and `norm_l2` with simd