--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Install toolchain
-        run: rustup toolchain install stable --profile minimal
+        run: rustup toolchain install 'nightly-2025-01-30' --profile minimal
 
       - uses: Swatinem/rust-cache@v2