--- a/phomo/src/lib.rs
+++ b/phomo/src/lib.rs
@@ -1,3 +1,4 @@
+#![feature(portable_simd)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![doc =include_str!("../README.md")]
 mod color_match;