… phomo-cli/tests/data/mosaic_transfer_tiles_to_master.png
… phomo/tests/data/match/matched.png
… phomo/tests/data/mosaic/master_matched_to_faces.png
… phomo/tests/data/mosaic/mosaic_16_16_match_master_to_tiles.png
… phomo/tests/data/mosaic/mosaic_16_16_match_tiles_to_master.png