Uses SIMD to squeeze a lot more performance out of the `norm_l{1,2}` metrics.

Some benchmarks to compare the before and after:
- before:
```console
❯ cargo bench metrics/norm
metrics/norm_l1         time:   [94.498 µs 94.671 µs 94.861 
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe

metrics/norm_l2         time:   [113.34 µs 113.84 µs 114.37 µs]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

- And with SIMD:

```console
❯ cargo bench metrics/norm
metrics/norm_l1         time:   [15.379 µs 15.420 µs 15.471 µs]
                        change: [-83.715% -83.665% -83.609%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

metrics/norm_l2         time:   [12.363 µs 12.388 µs 12.414 µs]
                        change: [-89.102% -89.036% -88.978%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
```

It requires rust nightly for the `portable_simd` feature.

I think switching to rust nightly changed the results of the color matching for some reason, so I updated the test images and I was getting some inconsistencies when running CI on the gh servers so I relaxed some test conditions.

This supersedes #23.

I'm sure SIMD can also be used to optimize the other metrics and probably the solvers.