@jonahclarsen 2025-02-07
With compiler optimization level 3, this actually made the distance_matrix calculation about 65% _slower_. I'm guessing the compiler already did some SIMD-type optimizations that were even more effective than your changes.