@loiccoyle 2025-02-07
hmm I'm not able to reproduce, as far as I know the `bench` profile inherits the `release` profile which by [default](https://doc.rust-lang.org/cargo/reference/profiles.html#release) is `opt-level=3`. So the benchmarks I ran should already be with `opt-level=3`.

How exactly did you find a 65% decrease? What benchmark did you run?

Here are some more benchmarks I did to sanity check:

```console
❯ cargo bench --no-default-features distance
distance_matrix         time:   [89.990 ms 90.171 ms 90.466 ms]
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
```
```console
❯ cargo bench --no-default-features --features=simd distance
distance_matrix         time:   [15.148 ms 15.169 ms 15.191 ms]
                        change: [-83.236% -83.178% -83.135%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
```
```console
❯ cargo bench --no-default-features --features=simd,parallel distance
Benchmarking distance_matrix: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 10.0s. You may wish to increase target time to 14.0s, enable flat sampling, or reduce sample count to 50.
distance_matrix         time:   [2.7767 ms 2.7847 ms 2.7928 ms]
                        change: [-81.724% -81.663% -81.591%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
```

And one with explicit `opt-level=3`:
```console
❯ RUSTFLAGS="-Copt-level=3" cargo bench --no-default-features distance
     Running benches/main.rs (/home/lcoyle/projects/phomo-rs/target/release/deps/main-ff9259d5c445ccd4)
distance_matrix         time:   [91.966 ms 92.222 ms 92.474 ms]
                        change: [+3201.0% +3215.5% +3228.8%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```
Which gives roughly the same as the `cargo bench --no-default-features distance` benchmark, as expected.