@loiccoyle 2025-02-07
As far as I know, the compiler won't do any simd unless `target-cpu=native` is set, it will then use the simd instruction set of the host cpu. Here are some benchmarks comparing `target-cpu=native` with the manual simd:

```console
❯ RUSTFLAGS="-Ctarget-cpu=native" cargo bench --no-default-features distance
Running benches/main.rs (/home/lcoyle/projects/phomo-rs/target/release/deps/main-ab427110c8cec6b1)
distance_matrix         time:   [26.671 ms 26.734 ms 26.801 ms]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

```console
❯ cargo bench --no-default-features --features=simd distance
distance_matrix         time:   [15.576 ms 15.616 ms 15.659 ms]
                        change: [-41.782% -41.588% -41.379%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 23 outliers among 100 measurements (23.00%)
  10 (10.00%) low mild
  8 (8.00%) high mild
  5 (5.00%) high severe
```

So from my benchmarks manually implementing simd leads to more gains than relying on the compiler.