@jonahclarsen 2025-02-20
Hey sorry, I haven't had time to work on this for a couple weeks, hopefully I can get to it soon. I just used rust's time to measure how long that part took across a few trials that I ran manually, it wasn't very robust. I will investigate further when I have more time.